REL-4333 Apply patch from Oracle to fix the bug: CPAN RT#121873
Patch is taken unchanged from Oracle's perl-GSSAPI RPM.
The patch is required because the module is very old, and latest
version 0.28 had a bug listed on CPAN RT with id #121873.
RPM vendors have applied the patch to fix this bug, so we also
bring this patch back for this reason.